+++ GUI/tclGUI/crackNum.tcl
@@ -44,6 +44,7 @@ set FORMAT_SECTIONS {
         {fe4m3    "FP8 (E4M3)"  fixed    e4m3}
         {fe5m2    "FP8 (E5M2)"  fixed    e5m2}
         {fe8m0    "FP8 (E8M0)"  fixed    e8m0}
+        {fue5m3   "FP8 (UE5M3)" fixed    ue5m3}
         {fbp      "Brain"       fixed    bp}
         {ftf32    "TF32"        fixed    tf32}
     }}
@@ -602,6 +603,7 @@ proc parseArgs {argv} {
                 fp4     { set state(selection) ffp4 }
                 fp4e0m3 { set state(selection) ffp4e0m3 }
                 e8m0    { set state(selection) fe8m0 }
+                ue5m3   { set state(selection) fue5m3 }
                 default {
                     if {[regexp {^(\d+)\+(\d+)$} $v _ e s]} {
                         set state(selection) fcs